third_party/blink/web_tests/editing/selection/addRange-failures-expected.txt

CONSOLE WARNING: line 29: addRange(): The given range isn't in document.
CONSOLE WARNING: line 46: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Test error handling of Selection.addRange().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Running: testNull
PASS selection.addRange(null) threw exception TypeError: Failed to execute 'addRange' on 'Selection': parameter 1 is not of type 'Range'..
PASS selection.rangeCount is 1
PASS selection.getRangeAt(0).startContainer === document.body is true
PASS selection.getRangeAt(0).startOffset is 0
PASS selection.getRangeAt(0).endContainer === document.body is true
PASS selection.getRangeAt(0).endOffset is 0
Running: testEmptyRange
PASS selection.rangeCount is 1
PASS selection.getRangeAt(0).startContainer === document.body is true
PASS selection.getRangeAt(0).startOffset is 0
PASS selection.getRangeAt(0).endContainer === document.body is true
PASS selection.getRangeAt(0).endOffset is 0
Running: testNotInDocumentRange
PASS selection.rangeCount is 1
PASS selection.getRangeAt(0).startContainer === document.body is true
PASS selection.getRangeAt(0).startOffset is 0
PASS selection.getRangeAt(0).endContainer === document.body is true
PASS selection.getRangeAt(0).endOffset is 0
Running: testOtherDocument
PASS selection.rangeCount is 1
PASS selection.getRangeAt(0).startContainer === document.body is true
PASS selection.getRangeAt(0).startOffset is 0
PASS selection.getRangeAt(0).endContainer === document.body is true
PASS selection.getRangeAt(0).endOffset is 0
Running: testOtherDocumentFragments
PASS selection.rangeCount is 1
PASS selection.getRangeAt(0).startContainer === document.body is true
PASS selection.getRangeAt(0).startOffset is 0
PASS selection.getRangeAt(0).endContainer === document.body is true
PASS selection.getRangeAt(0).endOffset is 0
PASS successfullyParsed is true

TEST COMPLETE
